requirements.txt
streamlit
python-dotenv
pandas

